# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DAYANERA Core API",
    "description": "Özel, yalnızca yerel (127.0.0.1) DAYANERA.ai çekirdek API'si. Tarayıcı, gelecekteki Windows ve Android istemcileri bu tek arayüzü kullanır. Herkese açık bir ürün API'si değildir.",
    "version": "0.1.0-beta"
  },
  "paths": {
...
    "/api/v1/messages/{message_id}/calculation": {
      "get": {
        "tags": [
          "conversations"
        ],
        "summary": "Ayrıntılı çözüm (hesap izi)",
        "operationId": "message_calculation_api_v1_messages__message_id__calculation_get",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Message Calculation Api V1 Messages  Message Id  Calculation Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/calculations": {
      "post": {
        "tags": [
          "calculations"
        ],
        "summary": "Deterministik hesap çalıştır",
        "operationId": "run_api_v1_calculations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalcIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalcOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "calculations"
        ],
        "summary": "Hesap geçmişi",
        "operationId": "history_api_v1_calculations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalcOut"
                  },
                  "title": "Response History Api V1 Calculations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calculations/{calc_id}": {
      "get": {
        "tags": [
          "calculations"
        ],
        "summary": "Hesap ayrıntısı",
        "operationId": "get_calc_api_v1_calculations__calc_id__get",
        "parameters": [
          {
            "name": "calc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalcOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalcIn": {
        "properties": {
          "calc_type": {
            "type": "string",
            "maxLength": 80,
            "title": "Calc Type"
          },
          "inputs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CalcInputIn"
            },
            "type": "object",
            "title": "Inputs"
          },
          "compare_with_llm": {
            "type": "boolean",
            "title": "Compare With Llm",
            "default": false
          },
          "llm_draft_outputs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Draft Outputs",
            "description": "Test/entegrasyon: harici bir taslak sonucu motorla karşılaştırmak için."
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          }
        },
        "type": "object",
        "required": [
          "calc_type",
          "inputs"
        ],
        "title": "CalcIn"
      },
      "CalcInputIn": {
        "properties": {
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "maxLength": 20,
            "title": "Unit",
            "default": ""
          },
          "provenance": {
            "type": "string",
            "enum": [
              "user_input",
              "extracted_value",
              "memory_item"
            ],
            "title": "Provenance",
            "default": "user_input"
          },
          "ref_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Id"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "CalcInputIn"
      },
      "CalcOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "calc_type": {
            "type": "string",
            "title": "Calc Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "engine_version": {
            "type": "string",
            "title": "Engine Version"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputs"
          },
          "llm_draft": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Draft"
          },
          "comparison": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparison"
          },
          "mismatch": {
            "type": "boolean",
            "title": "Mismatch"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "calc_type",
          "status",
          "engine_version",
          "result",
          "inputs",
          "llm_draft",
          "comparison",
          "mismatch",
          "created_at"
        ],
        "title": "CalcOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}